created pair and connection menu components
